[SPARK-54136][SQL] Extract plan merging logic from `MergeScalarSubqueries` to `PlanMerger`
